refactor(anyharness): split workspace lifecycle responsibilities
