+++ anyharness/crates/anyharness-lib/src/adapters/git/operations/mod.rs
@@ -10,3 +10,4 @@ pub mod revert_patches;
 pub mod staging;
 pub mod status;
 mod status_operation;
+pub mod worktrees;
