+++ anyharness/crates/anyharness-lib/src/workspaces/runtime/access.rs
@@ -1,0 +1,45 @@
+use super::WorkspaceRuntime;
+use crate::workspaces::model::WorkspaceRecord;
+
+impl WorkspaceRuntime {
+    pub fn set_lifecycle_cleanup_state(
+        &self,
+        workspace_id: &str,
+        lifecycle_state: &str,
+        cleanup_state: &str,
+        cleanup_operation: Option<&str>,
+        cleanup_error_message: Option<&str>,
+        cleanup_failed_at: Option<&str>,
+        cleanup_attempted_at: Option<&str>,
+    ) -> anyhow::Result<Option<WorkspaceRecord>> {
+        let now = chrono::Utc::now().to_rfc3339();
+        self.store.update_lifecycle_cleanup_state(
+            workspace_id,
+            lifecycle_state,
+            cleanup_state,
+            cleanup_operation,
+            cleanup_error_message,
+            cleanup_failed_at,
+            cleanup_attempted_at,
+            &now,
+        )?;
+        self.get_workspace(workspace_id)
+    }
+
+    pub fn find_active_workspace_by_path_and_kind(
+        &self,
+        path: &str,
+        kind: &str,
+    ) -> anyhow::Result<Option<WorkspaceRecord>> {
+        self.store.find_active_by_path_and_kind(path, kind)
+    }
+
+    pub fn find_active_worktree_by_path_excluding_id(
+        &self,
+        path: &str,
+        excluded_id: &str,
+    ) -> anyhow::Result<Option<WorkspaceRecord>> {
+        self.store
+            .find_active_by_path_and_kind_excluding_id(path, "worktree", excluded_id)
+    }
+}
