+++ scripts/check_anyharness_old_paths.py
@@ -10,6 +10,9 @@ BLOCKED_PATHS = [
     "anyharness/crates/anyharness-lib/src/sessions/mcp.rs",
     "anyharness/crates/anyharness-lib/src/sessions/runtime.rs",
     "anyharness/crates/anyharness-lib/src/sessions/store.rs",
+    "anyharness/crates/anyharness-lib/src/workspaces/runtime.rs",
+    "anyharness/crates/anyharness-lib/src/workspaces/service.rs",
+    "anyharness/crates/anyharness-lib/src/workspaces/store.rs",
     "anyharness/crates/anyharness-lib/src/adapters/git/branch_base.rs",
     "anyharness/crates/anyharness-lib/src/adapters/git/diff.rs",
     "anyharness/crates/anyharness-lib/src/adapters/git/operation.rs",
