+++ scripts/max_lines_allowlist.txt
@@ -23,9 +23,6 @@ anyharness/crates/anyharness-lib/src/domains/runtime_config/service.rs 1057 exis
 anyharness/crates/anyharness-lib/src/sessions/live_config.rs 675 existing AnyHarness oversized file
 anyharness/crates/anyharness-lib/src/sessions/prompt.rs 1423 existing AnyHarness oversized file
 anyharness/crates/anyharness-lib/src/sessions/service.rs 645 existing AnyHarness oversized file
-anyharness/crates/anyharness-lib/src/workspaces/runtime.rs 1844 existing AnyHarness oversized file
-anyharness/crates/anyharness-lib/src/workspaces/service.rs 952 existing AnyHarness oversized file
-anyharness/crates/anyharness-lib/src/workspaces/store.rs 734 existing AnyHarness oversized file
 anyharness/crates/proliferate-worker/src/commands/dispatcher.rs 1718 existing Proliferate worker oversized file
 anyharness/crates/proliferate-worker/src/commands/mapping.rs 954 existing Proliferate worker oversized file
 anyharness/sdk-react/src/hooks/sessions.ts 653 existing SDK React oversized file
